GameOver Screen / power Up 2 / Enemy Spawn Rate / ottimizzazione 2
In Scena:
 - Allontanata la telecamera dalla scena principale in modo tale che i pulsanti non si sovrappongono;
 - Spostati e Scalati gli oggetti di enviroment per compensare l'allontanamento della telecamera;
 - Impostati i font in grassetto nelle UI e nei Menu;

 - Spostata lo spawner dei nemici all'interno della relativa base,

 - Aggiunta Scermata di GameOver con pulsanti di "Retry" e "Main Menu";

Scripts:
 - Commentati script in generale;

 - Enemy.cs:
   - inserito calcolo per l'aumento di danno basato sul totale presente nello script "EnemySpawner.cs",
   - aggiunto counter quando muore per segnare i nemici uccisi;

 - EnemySpawner.cs:
   - Aggiunte variabili e funzioni annesse per calcolare il cambio di Rate e l'aumento del danno;

 - MenuManager.cs:
   - Eliminato il richiamo dell'EventAction per i soldi e spostato in UIManager per ottimizzare gli script;
   - Spostato l'attivazione dello Screen di gameOver in MenuManager;
   - Aggiunto kill Counter nella per la schermata di GameOver;
   - Aggiunta funzione per far tornare al Main menu nella schermata di GameOver;

 - UIManger.cs:
   - Spostato tutte le cose per far funzionare il Money Counter in questo script;

 - Turrets.cs:
   - Implementata funzione per interagire con il potenziamento delle torrette solo quando hai abbastanza soldi;

+++ Assets/Scripts/Interfaces/IDamageable.cs
@@ -1,5 +1,8 @@
 public interface IDamageable
 {
-    public void TakeDamage(int damage);
+    //Function used to deal Damage to the hittables (Enemy and Player) based on the Damage comunicated by the Attackers (Bullets and Enemy)
+    public void TakeDamage(float damage);
+
+    //Function used to Destroy the hittables when the Attackers finds out that the health is finished
     public void Despawn();
 }
